init Register
Huyph20687 đã khởi tạo màn đăng ký
