+++ Screen/Register.js
@@ -1,0 +1,19 @@
+import React,{Component} from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+export default class Register extends Component{
+    render(){
+        return (
+        <View style={styles.container}>
+            <Text>Register</Text>
+        </View>
+        );
+    }
+    }
+    const styles = StyleSheet.create({
+        container: {
+            flex: 1,
+            backgroundColor: '#fff',
+            alignItems: 'center',
+            justifyContent:'center',
+        },
+    });
